make submit invoice use its own hoc
