+++ src/components/snew/SubmitPage/SubmitPageProposal.js
@@ -1,22 +1,19 @@
-import React, { useState } from "react";
+import React from "react";
 import ReactBody from "react-body";
-import submitFormHOC from "../../hocs/submitForm";
-import MarkdownEditorField from "../Form/Fields/MarkdownEditorField";
-import FilesField from "../Form/Fields/FilesField";
-import ErrorField from "../Form/Fields/ErrorField";
-import SelectField from "../Form/Fields/SelectField";
-import InputFieldWithError from "../Form/Fields/InputFieldWithError";
-import Message from "../Message";
-import MultipleItemsBodyMessage from "../MultipleItemsBodyMessage";
+import MarkdownEditorField from "../../Form/Fields/MarkdownEditorField";
+import FilesField from "../../Form/Fields/FilesField";
+import ErrorField from "../../Form/Fields/ErrorField";
+import InputFieldWithError from "../../Form/Fields/InputFieldWithError";
+import Message from "../../Message";
+import MultipleItemsBodyMessage from "../../MultipleItemsBodyMessage";
 import isArray from "lodash/isArray";
 import isUndefined from "lodash/isUndefined";
 import concat from "lodash/concat";
 import cloneDeep from "lodash/cloneDeep";
 import { Field } from "redux-form";
-import MarkdownHelp from "../MarkdownHelp";
-import { MANAGE_CREDITS_MODAL } from "../Modal/modalTypes";
-import { PROPOSAL_GUIDELINES } from "../../constants";
-import InvoiceDatasheet from "../InvoiceDatasheet";
+import MarkdownHelp from "../../MarkdownHelp";
+import { MANAGE_CREDITS_MODAL } from "../../Modal/modalTypes";
+import { PROPOSAL_GUIDELINES } from "../../../constants";
 
 const normalizer = (value, previousValue) => {
   let files = [];
@@ -41,147 +38,6 @@ const normalizer = (value, previousValue) => {
 const ToggledMarkdown = props => (
   <MarkdownEditorField input={props.input} toggledStyle={true} />
 );
-
-const InvoiceSubmit = props => {
-  const {
-    isLoading,
-    PageLoadingIcon,
-    error,
-    warning,
-    onSave,
-    submitting,
-    handleSubmit,
-    validationError,
-    submitError,
-    editingMode,
-    valid
-  } = props;
-
-  const [datasheetErrors, setDatasheetErrors] = useState([]);
-  const submitEnabled = !submitting && valid && datasheetErrors.length === 0;
-
-  return isLoading ? (
-    <PageLoadingIcon />
-  ) : (
-    <div className="content" role="main">
-      <div className="page submit-proposal-page">
-        <ReactBody className="submit-page" />
-        <div className="submit conztent warn-on-unload" id="newlink">
-          {validationError && (
-            <Message type="error" header="Error creating invoice">
-              <MultipleItemsBodyMessage items={validationError} />
-            </Message>
-          )}
-          {error && (
-            <Message
-              type="error"
-              header="Error creating invoice"
-              body={error}
-            />
-          )}
-          {!error && warning && (
-            <Message type="warn" header="Warning">
-              <MultipleItemsBodyMessage items={warning} />
-            </Message>
-          )}
-          <div className="formtabs-content">
-            <div className="spacer">
-              <Field
-                name="global"
-                component={props => (
-                  <ErrorField title="Cannot submit invoice" {...props} />
-                )}
-              />
-              <div className="roundfield" id="title-field">
-                <div className="roundfield-content">
-                  <div style={{ display: "flex", width: "100%" }}>
-                    <Field
-                      name="month"
-                      component={SelectField}
-                      options={[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]}
-                      tabIndex={1}
-                      label="Month"
-                    />
-                    <Field
-                      name="year"
-                      component={SelectField}
-                      tabIndex={1}
-                      type="text"
-                      options={[2019, 2020, 2021, 2022, 2023, 2024, 2025]}
-                      label="Year"
-                    />
-                  </div>
-                  <div className="usertext">
-                    <Field
-                      name="name"
-                      label="Contractor name"
-                      type="text"
-                      component={InputFieldWithError}
-                    />
-                    <Field
-                      name="location"
-                      label="Contractor location"
-                      type="text"
-                      component={InputFieldWithError}
-                    />
-                    <Field
-                      name="contact"
-                      label="Contractor contact"
-                      type="text"
-                      component={InputFieldWithError}
-                    />
-                    <Field
-                      name="rate"
-                      label="Contractor rate"
-                      type="number"
-                      component={InputFieldWithError}
-                    />
-                    <Field
-                      name="address"
-                      label="Payment address"
-                      type="text"
-                      component={InputFieldWithError}
-                    />
-                  </div>
-                  <div className="usertext">
-                    <Field
-                      name="datasheet"
-                      onChangeErrors={setDatasheetErrors}
-                      errors={datasheetErrors}
-                      component={InvoiceDatasheet}
-                    />
-                    {/* <InvoiceDatasheet /> */}
-                  </div>
-                  <div className="submit-wrapper">
-                    <button
-                      className={`togglebutton access-required${!submitEnabled &&
-                        " not-active disabled"}`}
-                      name="submit"
-                      type="submit"
-                      value="form"
-                      onClick={handleSubmit(onSave)}
-                    >
-                      {!editingMode ? "submit" : "update"}
-                    </button>
-                  </div>
-                  {submitError ? (
-                    <Message
-                      type="error"
-                      header={`Error ${
-                        editingMode ? "updating" : "creating"
-                      } invoice`}
-                      body={submitError}
-                    />
-                  ) : null}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const ProposalSubmit = props => {
   const {
@@ -360,14 +216,16 @@ const ProposalSubmit = props => {
   );
 };
 
-class SubmitPage extends React.Component {
-  render() {
-    if (this.props.isCMS) {
-      return <InvoiceSubmit {...this.props} />;
-    } else {
-      return <ProposalSubmit {...this.props} />;
-    }
-  }
-}
+export default ProposalSubmit;
 
-export default submitFormHOC(SubmitPage);
+// class SubmitPage extends React.Component {
+//   render() {
+//     if (this.props.isCMS) {
+//       return <InvoiceSubmit {...this.props} />;
+//     } else {
+//       return <ProposalSubmit {...this.props} />;
+//     }
+//   }
+// }
+
+// export default submitFormHOC(SubmitPage);
