+++ src/connectors/submit.js
@@ -10,16 +10,11 @@ const submitConnector = connect(
     userid: sel.userid,
     username: sel.loggedInAsUsername,
     keyMismatch: sel.getKeyMismatch,
-    proposalCredits: sel.proposalCredits,
-    token: sel.newInvoiceToken,
-    invoiceError: sel.newInvoiceError
+    proposalCredits: sel.proposalCredits
   }),
   {
     onFetchData: act.onGetPolicy,
-    onSaveInvoice: act.onSaveNewInvoice,
-    onSaveProposal: act.onSaveNewProposal,
     openModal: act.openModal,
-    onResetInvoice: act.onResetInvoice,
     onSaveDraft: act.onSaveDraftProposal,
     onDeleteDraft: act.onDeleteDraftProposal
   }
